requirements.txt
scrapy
scrapy-selenium
selenium
pymongo
python-dotenv